requirements.txt
asgiref==3.8.1
cloudinary==1.40.0
Django==3.2.25
django-cloudinary-storage==0.3.0
django-filter==23.5
djangorestframework==3.14.0
Pillow==8.2.0
pytz==2024.1
sqlparse==0.5.0
